Refactor playlist

+++ src-tauri/src/interface/playlist/impls.rs
@@ -1,6 +1,8 @@
 use eyre::Result;
 use std::path::PathBuf;
 
+use crate::utils::sha1_hash;
+
 use super::{Playlist, PlaylistInfo};
 
 impl Playlist {
@@ -10,10 +12,23 @@ impl Playlist {
         Ok(Self {
             info: playlist,
             path: path.clone(),
+            hash: sha1_hash(&path.as_os_str().to_string_lossy()),
         })
     }
-    pub async fn save(&self) -> Result<()> {
+
+    pub fn with_path(playlist: PlaylistInfo, path: PathBuf) -> Self {
+        Self {
+            info: playlist,
+            hash: sha1_hash(&path.as_os_str().to_string_lossy()),
+            path,
+        }
+    }
+
+    pub async fn save(&self, allow_overwrite: bool) -> Result<()> {
         let playlist_str = serde_json::to_string_pretty(&self.info)?;
+        if !allow_overwrite && self.path.exists() {
+            return Err(eyre::eyre!("Playlist already exists"));
+        }
         tokio::fs::write(&self.path, playlist_str).await?;
         Ok(())
     }
